# docs/openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "operationId": "ProductsController_list",
        "summary": "List public products with filters",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search nameAr / nameEn / SKU (ILIKE)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isFeatured",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isBestSeller",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "inStock",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Admin filter: active products only",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "missingImages",
            "required": false,
            "in": "query",
            "description": "Admin filter: products with no images",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "lowStock",
            "required": false,
            "in": "query",
            "description": "Admin filter: low stock products",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Admin sort",
            "schema": {
              "enum": [
                "newest",
                "name",
                "price",
                "stock"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "summary": "Get a public product by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/v1/admin/products/{id}": {
      "get": {
        "operationId": "AdminProductsController_findOne",
        "summary": "Admin get product by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "AdminProductsController_update",
        "summary": "Update product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminProductsController_remove",
        "summary": "Soft-delete product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/reviews": {
      "get": {
        "operationId": "AdminReviewsController_list",
        "summary": "List product reviews (admin)",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/reviews/{id}/approve": {
      "patch": {
        "operationId": "AdminReviewsController_approve",
        "summary": "Approve a review and recompute product rating",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/reviews/{id}/reject": {
      "patch": {
        "operationId": "AdminReviewsController_reject",
        "summary": "Reject a review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/reviews/{id}/hide": {
      "patch": {
        "operationId": "AdminReviewsController_hide",
        "summary": "Hide an approved review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review hidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/reviews/{id}/show": {
      "patch": {
        "operationId": "AdminReviewsController_show",
        "summary": "Show an approved review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review shown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Tharaa Market API",
    "description": "Official **Version 1** REST API contract for Tharaa Market (Al Khafji).\n\n### Stability\nThis API is **frozen**. Breaking changes to URLs, request/response bodies, enums, pagination, auth, uploads, or error envelopes are not allowed in v1.\nFuture incompatible changes must ship as `/api/v2`.\n\n### Response envelope\nSuccess: `{ success: true, data, meta? }`\nError: `{ success: false, statusCode, message, errorCode?, details?, timestamp, path }`\n\n### Authentication\nUse `Authorization: Bearer <accessToken>`. Obtain tokens via `/auth/login`, `/auth/register`, `/auth/google`, or `/auth/refresh`.\nRefresh tokens rotate on every use.\n\n### Uploads\nMultipart form field name: `file`. Allowed MIME: jpeg, png, webp. Served under `/static/`.\n\n### Admin realtime (Socket.IO)\nNamespace `/admin` for staff dashboards. Authenticate with the JWT access token (`auth.token` or `Authorization: Bearer`).\nEvents: `admin:order_created`, `admin:order_updated`. Roles: ADMIN, MANAGER, EMPLOYEE.\nFull contract: `docs/ADMIN_INTEGRATION.md` (Realtime section). REST remains the source of truth.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access JWT from auth endpoints"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false,
            "description": "Always false for errors"
          },
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "Validation failed",
            "description": "Human-readable error message"
          },
          "errorCode": {
            "type": "string",
            "example": "Bad Request",
            "description": "Optional machine-readable error code"
          },
          "details": {
            "type": "object",
            "description": "Optional extra details (e.g. validation messages)"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-14T00:00:00.000Z",
            "description": "ISO-8601 timestamp when the error was produced"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/products",
            "description": "Request path that produced the error"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ApiSuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "description": "Always true for success"
          },
          "data": {
            "type": "object",
            "description": "Payload returned by the endpoint"
          },
          "meta": {
            "type": "object",
            "description": "Optional pagination or extra metadata",
            "additionalProperties": true
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "nameAr": {
            "type": "string",
            "example": "تفاح"
          },
          "nameEn": {
            "type": "string",
            "example": "Apples"
          },
          "descriptionAr": {
            "type": "string"
          },
          "descriptionEn": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "unit": {
            "type": "string",
            "example": "kg"
          },
          "hasVariants": {
            "type": "boolean",
            "default": false
          },
          "regularPrice": {
            "type": "number",
            "example": 25.5
          },
          "salePrice": {
            "type": "number",
            "example": 19.99
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "isFeatured": {
            "type": "boolean",
            "default": false
          },
          "isBestSeller": {
            "type": "boolean",
            "default": false
          },
          "lowStockThreshold": {
            "type": "number",
            "default": 5
          },
          "initialQuantity": {
            "type": "number",
            "description": "Initial stock when product has no variants",
            "example": 0
          }
        }
      },
...
    }
  }
}